SCB-2094 Add SCManager API
