+++ datasource/etcd/engine_test.go
@@ -1,22 +1,25 @@
-// Licensed to the Apache Software Foundation (ASF) under one or more
-// contributor license agreements.  See the NOTICE file distributed with
-// this work for additional information regarding copyright ownership.
-// The ASF licenses this file to You under the Apache License, Version 2.0
-// (the "License"); you may not use this file except in compliance with
-// the License.  You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
-package task_test
+package etcd_test
 
 // initialize
 import (
+	"github.com/apache/servicecomb-service-center/datasource"
 	_ "github.com/apache/servicecomb-service-center/server/bootstrap"
 	"github.com/apache/servicecomb-service-center/server/core/proto"
 )
@@ -31,7 +34,6 @@ import (
 	pb "github.com/apache/servicecomb-service-center/pkg/registry"
 	"github.com/apache/servicecomb-service-center/pkg/util"
 	apt "github.com/apache/servicecomb-service-center/server/core"
-	"github.com/apache/servicecomb-service-center/server/task"
 	"github.com/astaxie/beego"
 
 	. "github.com/onsi/ginkgo"
@@ -47,7 +49,7 @@ var _ = BeforeSuite(func() {
 	beego.AppConfig.Set("registry_plugin", "etcd")
 	//clear service created in last test
 	time.Sleep(timeLimit)
-	task.ClearNoInstanceServices(timeLimit)
+	_ = datasource.Instance().ClearNoInstanceServices(context.Background(), timeLimit)
 })
 
 // map[domainProject][serviceName]*serviceCleanInfo
@@ -140,7 +142,7 @@ func serviceClearCheckFunc(domain string, project string) func() {
 			createService(domain, project, "svc3", withNoInstance, shouldNotClear)
 			createService(domain, project, "svc4", withInstance, shouldNotClear)
 
-			err = task.ClearNoInstanceServices(timeLimit)
+			err = datasource.Instance().ClearNoInstanceServices(context.Background(), timeLimit)
 			Expect(err).To(BeNil())
 
 			checkServiceCleared(domain, project)
